Rename all the modules
This renames the uploader module to a generic cli module in the
new pacifica namespace.
