+++ .github/ISSUE_TEMPLATE.md
@@ -1,4 +1,4 @@
-### Pacifica CLI Uploader version
+### Pacifica CLI Tool version
 [Version of the pacifica software where you are encountering the issue]
 
 ### Pacifica Core Software versions
