+++ .travis.yml
@@ -9,6 +9,10 @@ python:
  - 2.7
  - 3.6
 
+stages:
+- lint
+- test
+
 addons:
   apt:
     sources:
@@ -16,12 +20,23 @@ addons:
     packages:
       - elasticsearch
       - oracle-java8-set-default
-env:
-- RUN_LINTS=true
-- RUN_LINTS=
 
 cache: pip
 
 before_install: bash -xe travis/before-install.sh
 install: pip install -r requirements-dev.txt
-script: bash -xe travis/test-script.sh
+script:
+- bash -xe travis/unit-tests.sh
+- bash -xe travis/test-deploy.sh
+
+jobs:
+  include:
+  - stage: lint
+    python: 3.6
+    script:
+    - pre-commit run -a
+    - radon cc pacifica
+  - python: 2.7
+    script:
+    - pre-commit run -a
+    - radon cc pacifica
