--- CLIUploader.py
@@ -1,8 +1,0 @@
-#!/usr/bin/python
-# -*- coding: utf-8 -*-
-# pylint: disable=invalid-name
-"""Run the pacifica cli uploader."""
-from uploader_cli import main
-
-if __name__ == '__main__':
-    main()
