+++ pacifica/__init__.py
@@ -1,0 +1,4 @@
+#!/usr/bin/python
+# -*- coding: utf-8 -*-
+"""Pacifica Namespace."""
+__import__('pkg_resources').declare_namespace(__name__)
