+++ pacifica/cli/__init__.py
@@ -1,0 +1,3 @@
+#!/usr/bin/python
+# -*- coding: utf-8 -*-
+"""The main cli module."""
