+++ setup.py
@@ -5,22 +5,22 @@ try:  # pip version 9
     from pip.req import parse_requirements
 except ImportError:
     from pip._internal.req import parse_requirements
-from setuptools import setup
+from setuptools import setup, find_packages
 
 # parse_requirements() returns generator of pip.req.InstallRequirement objects
 INSTALL_REQS = parse_requirements('requirements.txt', session='hack')
 
 setup(
-    name='PacificaCLIUploader',
+    name='pacifica-cli',
     use_scm_version=True,
     setup_requires=['setuptools_scm'],
-    description='Pacifica CLI Uploader',
+    description='Pacifica CLI Tool',
     author='David Brown',
     author_email='david.brown@pnnl.gov',
-    packages=['uploader_cli'],
-    scripts=['CLIUploader.py'],
+    packages=find_packages(),
+    namespace_packages=['pacifica'],
     entry_points={
-        'console_scripts': ['CLIUploader=uploader_cli:main'],
+        'console_scripts': ['pacifica-cli=pacifica.cli.__main__:main'],
     },
     install_requires=[str(ir.req) for ir in INSTALL_REQS]
 )
